path: Implement arcs
This is an elliptical arc implementation according
to the SVG spec. The code is mostly taken from
librsvg, but pretty directly follows the SVG spec
implementation notes. This will be used in subsequent
commits.

We don't export this, since the parametrization
is inconvenient. We do want an arc_to API, but
these are not the arcs we are looking for.